Properly round the printed snr.
git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7917 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ lib/ft8_decode.f90
@@ -77,7 +77,7 @@ contains
         call timer('ft8b    ',0)
         call ft8b(dd,newdat,nfqso,ndepth,lsubtract,iaptype,icand,sync,f1,   &
              xdt,apsym,nharderrors,dmin,nbadcrc,iap,ipass,iera,message,xsnr)
-        nsnr=xsnr  
+        nsnr=nint(xsnr) 
         xdt=xdt-0.6
         call timer('ft8b    ',1)
         if(nbadcrc.eq.0 .and. associated(this%callback)) then
